everything is going fine

+++ src/components/Shop/Shop.js
@@ -15,16 +15,14 @@ const Shop = () => {
     const [item, setItem] = useState([]);
     const addToCart = (product2) => {
         const newItem = [...item, product2];
-        setItem(newItem);
         // let newCart = [];
-        // const exists = item.find(product => product.id === product2.id);
-        // if (!exists) {
-        //     newCart = [...item, product2];
-        // }
-        // else {
-        //     const rest = item.filter(product => product.id !== product2.id);
-        //     newCart = [...rest, exists]
-        // }
+        const exists = item.find(product => product.id === product2.id);
+        if (!exists) {
+            setItem(newItem);
+        }
+        else {
+            alert('Item Already Selected');
+        }
         // setItem(newCart);
     }
 
@@ -43,6 +41,8 @@ const Shop = () => {
                 <div className="cart-div">
                     <h2 className='cart-header'>Selected Items</h2>
                     <Cart item={item}></Cart>
+                    <button className='btn'>Choose One For Me</button><br />
+                    <button className='btn'>Remove All</button>
                 </div>
             </div>
         </div>
